# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/token/validate": {
            "get": {
                "description": "Validates your auth token from your browser cookie",
                "tags": [
                    "token"
                ],
                "summary": "Validate Token",
                "responses": {
                    "200": {
                        "description": "Authorized",
                        "schema": {
                            "type": ""
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No Cookie",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unknown internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Basic password login token auth, stored in a cookie for convenience",
            "type": "apiKey",
            "name": "Authorization",
            "in": "cookie"
        }
    }
}